[Refactor] Aws S3 파일 삭제, 기본이미지 설정

+++ src/main/java/com/bob/mate/domain/aws/service/S3Service.java
@@ -13,14 +13,18 @@ import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.util.Objects;
 
 @Component
 @RequiredArgsConstructor
 @Slf4j
 public class S3Service {
 
+    private static final String DEFAULT_IMAGE = "https://d3afymv2nzz1pw.cloudfront.net/doji.png";
+
     private final AmazonS3Client amazonS3Client;
     private final FileStore fileStore;
+
 
     @Value("${cloud.aws.s3.bucket}")
     public String bucket;
@@ -39,6 +43,11 @@ public class S3Service {
     }
 
     public void deleteImage(String storeFilename) {
-        amazonS3Client.deleteObject(bucket, storeFilename);
+        String filename = StringUtils.getFilename(storeFilename);
+        if (!Objects.equals(storeFilename, DEFAULT_IMAGE) && amazonS3Client.doesObjectExist(bucket, filename)) {
+            log.info("S3Uploader, S3에서 이미지(이미지명: {})를 삭제했습니다.", storeFilename);
+            amazonS3Client.deleteObject(bucket, storeFilename);
+        }
     }
+
 }
